pyproject.toml
[project]
name = "hiresignal"
version = "1.0.0"
description = "Agentic AI hiring intelligence pipeline"
requires-python = ">=3.11"

dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-tavily>=0.1.0",
    "langgraph>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "streamlit>=1.40.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "tools", "utils", "prompt_library"]
